Moved all styles to separate files.
