+++ src/screens/ForgottenPasswordScreen.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import { StyleSheet, View, Text, TouchableHighlight } from 'react-native';
+import { View, Text, TouchableHighlight } from 'react-native';
+import forgottenPasswordStyles from './style/forgottenPasswordStyles';
 
 export default class ForgottenPasswordScreen extends React.Component {
 
@@ -9,30 +10,11 @@ export default class ForgottenPasswordScreen extends React.Component {
 
   render() {
       return (
-          <View style={styles.container}>
-              <TouchableHighlight style={styles.button} onPress={() => this.handleButtonPress()}>
-                  <Text style={styles.text}>Forgotten Password Screen</Text>
+          <View style={forgottenPasswordStyles.container}>
+              <TouchableHighlight style={forgottenPasswordStyles.button} onPress={() => this.handleButtonPress()}>
+                  <Text style={forgottenPasswordStyles.text}>Forgotten Password Screen</Text>
               </TouchableHighlight>
           </View>
       );
   }
 }
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        backgroundColor: '#fff',
-        alignItems: 'center',
-        justifyContent: 'center',
-    },
-    button: {
-        backgroundColor: '#4F6D7A',
-        width: 250,
-        height: 50,
-        justifyContent: 'center',
-    },
-    text: {
-        textAlign: 'center',
-        color: '#fff',
-    }
-});
