+++ src/screens/style/newReportStyles.js
@@ -1,0 +1,19 @@
+import { Platform, StyleSheet } from 'react-native';
+
+const newReportStyles = StyleSheet.create({
+    MainContainer: {
+        justifyContent: 'center',
+        flex: 1,
+        paddingTop: (Platform.OS === 'ios') ? 20 : 0,
+    },
+
+    TextInputStyleClass: {
+        textAlign: 'center',
+        marginBottom: 7,
+        height: 40,
+        borderWidth: 1,
+        borderColor: '#FF5722',
+    },
+});
+
+export default  newReportStyles;
