+++ src/screens/style/signUpStyles.js
@@ -1,0 +1,22 @@
+import { StyleSheet } from 'react-native';
+
+const signUpStyles = StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: '#fff',
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    button: {
+        backgroundColor: '#4F6D7A',
+        width: 250,
+        height: 50,
+        justifyContent: 'center',
+    },
+    text: {
+        textAlign: 'center',
+        color: '#fff',
+    }
+});
+
+export default signUpStyles;
